Customized version of nfreader to index IP addresses
This file should be copied in the bin directory of nfdump
tool by Peter Haag and manually compiled as it is stated in
the header.

How the index works:

In redis keys are created such as n4:10.0.0.1. Each key has a set
as value. An IP address always starts with n followed by 4 (IPv4)
or by 6 (IPv6). This prefix is followed by the string representation
of the IP address. The associated set contains numbers such as 1,2,3...
where each number is the index of a file name. A set can be
queried in redis with the SMEMBERS command. The associated filename
can be queried with the command GET d:x where x is the index (number)
in the queried set.