Store only the basename of the file
Sometimes the nfcapd files are in various directories. However,
the nfcapd files include time stamps and for one collector the are
unique. If an nfcapd file is moved in another directory and then
indexed, then a new file index would be created. This is avoided
by using basename because then only the filename is stored.