Replicate created nfcapd files to an other host
Nfcapd can be executed with the command line option -x redis-cli rpush
toprocess %d/%f. Each time a nfcapd file is ready the file name is pushed in
a redis queue toprocess. The nfdump-replicator fetches then these files and
transfers them to another host. If the transfer was successfull the filename
is put in a queue to delete.